backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
mysql-connector-python==8.3.0
openai==1.65.3
google-generativeai==0.8.0
google-auth==2.28.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0
pytesseract==0.3.10
bcrypt==4.1.2
